requirements.txt
fastapi
uvicorn
python-multipart
python-jose
sqlmodel
python-dotenv
pyodbc
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0